app/image_processor/requirements.txt
fastapi
uvicorn[standard]
pydantic
imagededup
python-multipart
Pillow
starlette
httpx
geopy